# apps/core-api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers": {
      "post": {
        "operationId": "CustomerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Customer"
        ]
      },
      "get": {
        "operationId": "CustomerController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PRIVATE",
                "COMPANY"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CustomerResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Customer"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "operationId": "CustomerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "historyPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "historyLimit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Customer"
        ]
      },
      "patch": {
        "operationId": "CustomerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Customer"
        ]
      },
      "delete": {
        "operationId": "CustomerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customer"
        ]
      }
    },
...
    "/api/employees": {
      "get": {
        "operationId": "EmployeeController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EmployeeRole"
            }
          },
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include inactive employees when true",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeesListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      },
      "post": {
        "operationId": "EmployeeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      }
    },
    "/api/employees/{id}": {
      "get": {
        "operationId": "EmployeeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      },
      "patch": {
        "operationId": "EmployeeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "employees"
        ]
      },
      "delete": {
        "operationId": "EmployeeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDeleteResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Employee is referenced by workshop orders"
          }
        },
        "tags": [
          "employees"
        ]
      }
    },
...
  },
  "info": {
    "title": "Auto Core Platform API",
    "description": "Generated OpenAPI spec for contract checks and client types.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {}
      },
      "CustomerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "COMPANY"
            ]
          },
          "company_name": {
            "type": "string",
            "nullable": true
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address_city": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "first_name",
          "last_name",
          "createdAt",
          "updatedAt"
        ]
      },
      "VehicleSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "plate": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "make",
          "model",
          "year"
        ]
      },
      "SalesOrderSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "order_number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "CONFIRMED",
              "IN_PROGRESS",
              "COMPLETED",
              "INVOICED"
            ]
          },
          "total_amount": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_number",
          "status",
          "total_amount",
          "createdAt"
        ]
      },
      "WorkshopTaskLineItemSummaryDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "string"
          },
          "unit_price": {
            "type": "string"
          }
        },
        "required": [
          "quantity",
          "unit_price"
        ]
      },
      "WorkshopTaskSummaryDto": {
        "type": "object",
        "properties": {
          "line_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopTaskLineItemSummaryDto"
            }
          }
        }
      },
      "WorkshopOrderSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "INTAKE",
              "IN_PROGRESS",
              "COMPLETED",
              "INVOICED"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "vehicle_id": {
            "type": "string",
            "nullable": true
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleSummaryDto"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopTaskSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "createdAt"
        ]
      },
      "InvoiceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoice_number": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "FINALIZED",
              "ISSUED",
              "PAID",
              "CANCELLED"
            ]
          },
          "date": {
            "type": "string"
          },
          "total_gross": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "date",
          "total_gross"
        ]
      },
      "CustomerHistoryMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "pageSize",
          "totalCount",
          "pageCount",
          "hasMore"
        ]
      },
      "CustomerDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "COMPANY"
            ]
          },
          "company_name": {
            "type": "string",
            "nullable": true
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "vat_id": {
            "type": "string",
            "nullable": true
          },
          "address_street": {
            "type": "string",
            "nullable": true
          },
          "address_city": {
            "type": "string",
            "nullable": true
          },
          "address_zip": {
            "type": "string",
            "nullable": true
          },
          "address_country": {
            "type": "string",
            "nullable": true
          },
          "vehicles": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleSummaryDto"
            }
          },
          "sales_orders": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesOrderSummaryDto"
            }
          },
          "workshop_orders": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopOrderSummaryDto"
            }
          },
          "invoices": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceSummaryDto"
            }
          },
          "workshop_orders_meta": {
            "$ref": "#/components/schemas/CustomerHistoryMetaDto"
          },
          "invoices_meta": {
            "$ref": "#/components/schemas/CustomerHistoryMetaDto"
          }
        },
        "required": [
          "id",
          "type",
          "first_name",
          "last_name",
          "workshop_orders_meta",
          "invoices_meta"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {}
      },
...
      "EmployeeRole": {
        "type": "string",
        "enum": [
          "MECHANIC",
          "SERVICE_ADVISOR",
          "PARTS_CLERK"
        ]
      },
      "EmployeeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmployeeRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "EmployeesListMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "EmployeesListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/EmployeesListMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmployeeRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "role"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmployeeRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "EmployeeDeleteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}